Added keyword parameter for general collection searching. Put collection tests into a separate file.
