+++ tests/test_granule.py
@@ -1,7 +1,7 @@
 import unittest
 
 from datetime import datetime
-from cmr.queries import GranuleQuery, CollectionQuery
+from cmr.queries import GranuleQuery
 
 class TestGranuleClass(unittest.TestCase):
 
@@ -341,12 +341,5 @@ class TestGranuleClass(unittest.TestCase):
         self.assertEqual(hits, 3)
 
 
-class TestCollectionClass(unittest.TestCase):
 
-    def test_archive_center(self):
-        query = CollectionQuery()
-        query.archive_center("LP DAAC")
-
-        self.assertIn("archive_center", query.params)
-        self.assertEqual(query.params["archive_center"], "LP DAAC")
